requirements.txt
DrissionPage==3.0.0
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0